Address review: split watchers out of test_misc.py
